refactored get_user into app.hindsite.common_model
